Alter word/char counting logic
The main logic to count the words remains the same:
if a non-word bound is followed by a non-CJK character, it counts it as a word.

However, if a character is a CJK character, it counts it as a char instead of word.
The main problem with the previous logic was that unlike Chinese and
Japanese, a single Korean character is not a word, and since it was counted
as a word, the reading time was significantly differ from the actual
time.

This commit suggests a change to the latin word counting logic as well.
Previously, a link was counted as a single word. I suggest counting all
the words that consist a link. For example, `https://google.com` is
now a three-word-paragraph. Hopefully, this change will improve the
accuracy of result.